# perf: avoid unnecessary clone