--- a/package/origlang-compiler/src/type_check.rs
+++ b/package/origlang-compiler/src/type_check.rs
@@ -1,5 +1,6 @@
 pub mod error;
 
+use std::borrow::Cow;
 use std::cell::RefCell;
 use std::collections::hash_map::RandomState;
 use std::collections::{HashMap, VecDeque};
@@ -252,13 +253,13 @@ fn handle_atomic_pattern(
             }])
         }
         AtomicPattern::Tuple(tp) => {
-            desugar(tp.clone(), expr, checker)
+            desugar(tp.into(), expr, checker)
         }
     }
 }
 
 fn desugar(
-    outer_destruction: Vec<AtomicPattern>, rhs: TypedExpression, checker: &TypeChecker,
+    outer_destruction: Cow<'_, [AtomicPattern]>, rhs: TypedExpression, checker: &TypeChecker,
 ) -> Result<Vec<TypedStatement>, TypeCheckError> {
     debug!("check: {outer_destruction:?} = {rhs:?}");
 
@@ -279,7 +280,7 @@ fn desugar(
                     } else {
                         debug!("tuple arity mismatch");
                         Err(TypeCheckError::UnsatisfiablePattern {
-                            pattern: AtomicPattern::Tuple(outer_destruction),
+                            pattern: AtomicPattern::Tuple(outer_destruction.to_vec()),
                             expression: TypedExpression::Variable { ident, tp: Type::tuple(tuple_element_types.clone()) },
                             expr_type: Type::tuple(tuple_element_types.clone()),
                         })
@@ -288,7 +289,7 @@ fn desugar(
                 other => {
                     debug!("non-tuple expression");
                     Err(TypeCheckError::UnsatisfiablePattern {
-                        pattern: AtomicPattern::Tuple(outer_destruction),
+                        pattern: AtomicPattern::Tuple(outer_destruction.to_vec()),
                         expression: TypedExpression::Variable { ident, tp: other.clone() },
                         expr_type: other,
                     })
@@ -371,7 +372,7 @@ fn desugar(
         other => {
             debug!("unsupported expression");
             Err(TypeCheckError::UnsatisfiablePattern {
-                pattern: AtomicPattern::Tuple(outer_destruction),
+                pattern: AtomicPattern::Tuple(outer_destruction.to_vec()),
                 expr_type: other.actual_type(),
                 expression: other,
             })