address #280

## before

```
git -c advice.detachedHead=false checkout a8aeb89a8399acb5d7ced9a3d27cba1e87137445 && CARGO_PROFILE_RELEASE_DEBUG=true RUSTFLAGS="-Awarnings" cargo run -p origlang-cli --release -
- execute --input-file ./compile/positive/perf/nested_tuple_2_1000.origlang
HEAD is now at a8aeb89 Merge pull request #282 from KisaragiEffective/fix/testsuite/log-regression
    Finished release [optimized + debuginfo] target(s) in 0.02s
     Running `target/release/origlang-cli execute --input-file ./compile/positive/perf/nested_tuple_2_1000.origlang`
init: 46ns
source: 1.483764ms
parser.ctor: 3.162688ms
parser.parse: 1.234062656s
typeck.ctor: 1.234070976s
typeck.check: 10.915282914s
runtime.ctor: 10.915301411s
runtime.run: 10.955619282s
```

## after

```
git -c advice.detachedHead=false checkout 981b47f55401e4fc0bed7c130dfcea9a3ad705d0 && CARGO_PROFILE_RELEASE_DEBUG=true RUSTFLAGS="-Awarnings" cargo run -p origlang-cli --release -- execute --input-file ./compile/positive/perf/nested_tuple_2_1000.origlang
HEAD is now at 981b47f refactor: express intuition more explicitly
    Finished release [optimized + debuginfo] target(s) in 0.02s
     Running `target/release/origlang-cli execute --input-file ./compile/positive/perf/nested_tuple_2_1000.origlang`
init: 51ns
source: 1.448468ms
parser.ctor: 3.162982ms
parser.parse: 1.165646592s
typeck.ctor: 1.165660387s
typeck.check: 1.635197136s
runtime.ctor: 1.635209278s
runtime.run: 1.671664304s
```